Address review follow-ups for renumber and wrap
Apply Oxford spelling to the wrap CHANGELOG entry, document the
`renumber_footnotes` API and the `parsing` re-export shim, and add
rustdoc for the `pub(super)` types and functions in
`renumber::definitions` so cargo doc surfaces their contracts.

Extract the inline `mod tests` block from `definitions.rs` into a
sibling `definitions/tests.rs` so the implementation file stays focused.

Strengthen `handle_prefix_line_can_repeat_or_change_the_continuation_prefix`
with a `repeat_prefix = true` case, and add a join-roundtrip assertion to
`wrap_text_preserves_hyphenated_words` so accidental character loss is
caught directly.
